requirements.txt
click>=8.0.0
toml>=0.10.0
PyYAML>=6.0